requirements.txt
alembic==1.14.0
asyncpg==0.30.0
bcrypt==4.3.0
fastapi==0.115.13
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic[email]==2.11.7
pyjwt==2.10.1
python-multipart==0.0.20
sqlalchemy[asyncio]==2.0.36
uvicorn[standard]==0.34.3
